requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
av==17.0.0
certifi==2026.2.25
click==8.3.2
croniter==6.2.2
ctranslate2==4.7.1
fastapi==0.136.0
faster-whisper==1.2.1
filelock==3.28.0
flatbuffers==25.12.19
fsspec==2026.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.11.0
idna==3.11
imageio-ffmpeg==0.6.0
markdown-it-py==4.0.0
mdurl==0.1.2
mpmath==1.3.0
numpy==2.4.4
onnxruntime==1.24.4
packaging==26.1
protobuf==7.34.1
pydantic==2.13.2
pydantic_core==2.46.2
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
PyYAML==6.0.3
redis==7.4.0
rq==2.8.0
rich==15.0.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
starlette==1.0.0
sympy==1.14.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
